requirements.txt
asgiref==3.9.0
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
cffi==1.17.1
channels==4.2.2
channels_redis==4.2.1
constantly==23.10.4
cryptography==45.0.5
daphne==4.2.1
Django==5.2.4
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
msgpack==1.1.1
phonenumbers==9.0.8
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.1.0
redis==6.2.0
service-identity==24.2.0
setuptools==80.9.0
sqlparse==0.5.3
Twisted==25.5.0
txaio==25.6.1
typing_extensions==4.14.1
whitenoise==6.9.0
zope.interface==7.2
